Better error messages for Json and Binary readers with regards to type tags/types

+++ mgen-cpplib/src/main/cpp/mgen/serialization/BinaryReader.h
@@ -274,8 +274,8 @@ private:
             const BINARY_TAG tag = readTag();
             if (tag != expTag) {
                 throw UnexpectedTypeException(
-                        S("BinaryReader::verifyReadTagIf: Unexpected tag ").append(S((int )expTag)).append(
-                                " but got ").append(S((int )tag)));
+                        S("BinaryReader::verifyReadTagIf: Unexpected tag ").append(S(expTag)).append(" but got ").append(
+                                S(tag)));
             }
         }
     }
